+++ src/server/netease/web_api/request.rs
@@ -36,8 +36,8 @@ impl SearchTarget {
 pub async fn search(
     search_target: SearchTarget,
     content: &str,
-    page: u16,
-    limit: u16,
+    page: i64,
+    limit: i64,
 ) -> Result<String, anyhow::Error> {
     if page == 0 {
         return Err(anyhow::anyhow!("Page must be greater than 0"));
